SyncConfigs: handle patches and env vars on it's own
