+++ Source/Client/EarlyPatches/SettingsPatches.cs
@@ -1,10 +1,7 @@
 using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using System.Reflection;
 using HarmonyLib;
 using Multiplayer.Client.Patches;
-using Multiplayer.Client.Util;
 using Verse;
 
 namespace Multiplayer.Client.EarlyPatches
@@ -104,69 +101,5 @@ namespace Multiplayer.Client.EarlyPatches
         }
 
         static bool Prefix() => !TickPatch.Simulating;
-    }
-
-    // Affects both reading and writing
-    [EarlyPatch]
-    [HarmonyPatch(typeof(LoadedModManager), nameof(LoadedModManager.GetSettingsFilename))]
-    static class OverrideConfigsPatch
-    {
-        private static Dictionary<(string, string), ModContentPack> modCache = new();
-
-        static void Postfix(string modIdentifier, string modHandleName, ref string __result)
-        {
-            if (!Multiplayer.restartConfigs)
-                return;
-
-            if (!modCache.TryGetValue((modIdentifier, modHandleName), out var mod))
-            {
-                mod = modCache[(modIdentifier, modHandleName)] =
-                    LoadedModManager.RunningModsListForReading.FirstOrDefault(m =>
-                        m.FolderName == modIdentifier
-                        && m.assemblies.loadedAssemblies.Any(a => a.GetTypes().Any(t => t.Name == modHandleName))
-                    );
-            }
-
-            if (mod == null)
-                return;
-
-            if (JoinData.ignoredConfigsModIds.Contains(mod.ModMetaData.PackageIdNonUnique))
-                return;
-
-            __result = SyncConfigs.GetConfigPath(mod.PackageIdPlayerFacing.ToLowerInvariant(), modHandleName);
-        }
-    }
-
-    [EarlyPatch]
-    [HarmonyPatch]
-    static class HugsLib_OverrideConfigsPatch
-    {
-        public static string HugsLibConfigOverridenPath;
-
-        private static MethodInfo MethodToPatch = AccessTools.Method("HugsLib.Core.PersistentDataManager:GetSettingsFilePath");
-
-        static bool Prepare() => MethodToPatch != null;
-
-        static MethodInfo TargetMethod() => MethodToPatch;
-
-        static void Prefix(object __instance)
-        {
-            if (!Multiplayer.restartConfigs)
-                return;
-
-            if (__instance.GetType().Name != "ModSettingsManager")
-                return;
-
-            var newPath = Path.Combine(
-                GenFilePaths.FolderUnderSaveData(JoinData.TempConfigsDir),
-                GenText.SanitizeFilename($"{JoinData.HugsLibId}-{JoinData.HugsLibSettingsFile}")
-            );
-
-            if (File.Exists(newPath))
-            {
-                __instance.SetPropertyOrField("OverrideFilePath", newPath);
-                HugsLibConfigOverridenPath = newPath;
-            }
-        }
     }
 }
