+++ Source/Client/Windows/JoinDataWindow.cs
@@ -750,6 +750,7 @@ namespace Multiplayer.Client
             if (applyConfigs)
             {
                 SyncConfigs.SaveConfigs(data.remoteModConfigs);
+                SyncConfigs.MarkApplicableForChildProcess();
             }
 
             var connectTo = data.remoteSteamHost != null
@@ -758,7 +759,6 @@ namespace Multiplayer.Client
 
             // The env variables will get inherited by the child process started in GenCommandLine.Restart
             Environment.SetEnvironmentVariable(EarlyInit.RestartConnectVariable, connectTo);
-            Environment.SetEnvironmentVariable(EarlyInit.RestartConfigsVariable, applyConfigs ? "true" : "false");
 
             GenCommandLine.Restart();
         }
